fix: refactored the location of few kebab files
