+++ src/main/java/com/java/project/kebab/Kebab.java
@@ -1,4 +1,4 @@
-package com.java.project.kebab.domain;
+package com.java.project.kebab;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
