+++ src/main/java/com/java/project/kebab/KebabRepository.java
@@ -6,8 +6,6 @@ import java.util.List;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.java.project.kebab.domain.Kebab;
-
 @Repository
 public interface KebabRepository extends JpaRepository<Kebab,Integer>{
