+++ src/main/java/com/java/project/kebab/KebabService.java
@@ -2,9 +2,6 @@ package com.java.project.kebab;
 
 import java.util.List;
 
-import com.java.project.kebab.domain.KebabDTO;
-import com.java.project.kebab.payload.AddUpdateKebabRequest;
-
 public interface KebabService {
 
     List<KebabDTO> getAllKebabs();
